# internal/docs/openapi.yaml
openapi: 3.1.0
info:
  title: IAG Project Management API
  version: 1.0.0
  description: |
    Workspace-document API for the IAG Project Management service.
    Frontends should typically call through the api-gateway at
    `/api/v1/project-management/*`; the paths below are the suffix
    after that prefix is stripped.

    The companion narrative is `docs/FRONTEND_INTEGRATION.md`. Endpoint
    surface here is hand-maintained — keep both files in sync when
    adding or changing routes.
servers:
  - url: https://iag-api-gateway-production.up.railway.app/api/v1/project-management
    description: Production via gateway
  - url: http://localhost:8080/api/v1/project-management
    description: Local via gateway
  - url: http://localhost:4102
    description: Local direct (no gateway)
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    EntityID:
      in: path
      name: id
      required: true
      schema: { type: integer }
...
security:
  - bearer: []
paths:
...
  /public/forms/{slug}:
    get:
      tags: [forms]
      summary: Describe a public form (fields only, no auth)
      security: []
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses: { '200': { description: ok } }
    post:
      tags: [forms]
      summary: Submit a public form (creates a task in the configured project)
      security: []
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                values:
                  type: object
                  additionalProperties: { type: string }
              required: [values]
      responses: { '200': { description: ok } }
...
  /shared/{token}/task/{id}:
    get:
      tags: [share]
      summary: Read a single task via a public share token (no platform auth)
      security: []
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/EntityID'
      responses: { '200': { description: ok } }
  /shared/{token}/project/{id}:
    get:
      tags: [share]
      summary: Read a project (with its tasks) via a public share token
      security: []
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses: { '200': { description: ok } }
...